requirements.txt
torch
transformers
datasets
evaluate
Pillow
numpy
